Refresh bash and README, Makefile created
