+++ README.md
@@ -1,1 +1,10 @@
 # advanced_course_work
+
+В папке находится 4 файла: 
+- refsol.c - эталонное решение задачи с матрицами, считываемыми из файла testcase_1. Если ошибок в файле не найдено, создаёт три файла out_trans и out_multi, в которы записываются результаты работы. Также в консоль выводится ранг матриц, поток которого перенаправляется в создаваемый файл rang
+
+- generate.c - генератор исходных файлов с матрицами, на данный момент релиз неокончателен, поэтому стоит цикл на 1 раз, т.е. на создание одого файла - testcase_1. Цикл реализован в баш-файле
+
+- run.sh - запускает генератор тестовых заданий и программу, выдающую эталонный результат, ни с чем не сраванивает, затем удаляет появившиеся в процессе работы файлы (out_trans,out_multi и rang). В случае неправильных матриц пытается удалить несозданные файлы out_trans, out_multi и rang, но сообщение отключены.
+
+-Makefile
