--- README.txt
@@ -1,6 +1,0 @@
-В папке находится 3 файла: 
-- sol.c - работа с матрицами, считываемыми из файла testcase_1. Если ошибок в файле не найдено, создаёт два файла out_trans  и out_multi, в которы записываются результаты работы. Также в консоль выводится ранг матриц.
-
-- gen.c - генератор исходных файлов с матрицами, на данный момент релиз неокончателен, поэтому стоит цикл на 1 раз, т.е. на создание одого файла - testcase_1
-
-- run.sh - компилирует оба файла, запускает их, ни с чем не сраванивает, затем удаляет появившиеся файлы. В случае неправильных матриц пытается удалить несозданные файлы out_trans  и out_multi, что ведет к выводу ошибки на консоль.
